# api-gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_handler.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: study.khoipn.id.vn
info:
  contact:
    email: support@study.khoipn.id.vn
    name: Study Platform Support
    url: https://study.khoipn.id.vn/support
  description: API Gateway for Study Platform - A comprehensive learning management
    system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://study.khoipn.id.vn/terms
  title: Study Platform API
  version: "1.0"
paths:
  /analytics/user:
    get:
      description: Get learning analytics and statistics for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User analytics
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user analytics
      tags:
      - Analytics
...
  /enrollments:
    get:
      description: Get all enrollments for the authenticated user
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: page_size
        type: integer
      - description: Filter by status (active, completed, dropped)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of enrollments
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: List enrollments
      tags:
      - Enrollments
    post:
      consumes:
      - application/json
      description: Enroll user in a course
      produces:
      - application/json
      responses:
        "201":
          description: Enrolled successfully
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Create enrollment
      tags:
      - Enrollments
  /enrollments/courses/{course_id}:
    get:
      description: Get enrollment for a specific course
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Enrollment details
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
        "404":
          description: Enrollment not found
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Get enrollment details
      tags:
      - Enrollments
...
  /progress/courses/{course_id}:
    get:
      description: Get all progress for a course
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Course progress
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user course progress
      tags:
      - Progress
  /progress/courses/{course_id}/completion:
    get:
      description: Get completion percentage and statistics for a course
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Course completion stats
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Get course completion
      tags:
      - Progress
  /progress/courses/{course_id}/lectures/{lecture_id}:
    get:
      description: Get user's progress for a specific lecture
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: string
      - description: Lecture ID
        in: path
        name: lecture_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Progress details
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
        "404":
          description: Progress not found
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Get lecture progress
      tags:
      - Progress
  /progress/lectures/{lecture_id}:
    get:
      description: Get detailed progress information for a lecture
      parameters:
      - description: Lecture ID
        in: path
        name: lecture_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lecture progress
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
        "404":
          description: Progress not found
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Get lecture progress details
      tags:
      - Progress
  /progress/lectures/complete:
    post:
      consumes:
      - application/json
      description: Mark a lecture as completed
      produces:
      - application/json
      responses:
        "200":
          description: Lecture marked complete
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Mark lecture complete
      tags:
      - Progress
  /progress/update:
    post:
      consumes:
      - application/json
      description: Update user's learning progress for a lecture
      produces:
      - application/json
      responses:
        "200":
          description: Progress updated
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Update progress
      tags:
      - Progress
produces:
- application/json
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...